+++ test/unit/webdriver/webdriver_test.py
@@ -15,7 +15,6 @@
 import unittest
 import httpretty
 import json
-import sys
 
 from appium import webdriver
 from test.unit.helper.test_helper import TestHelper
@@ -85,20 +84,6 @@ class WebDriverWebDriverTests(unittest.TestCase):
 
         self.assertEqual('session-id', driver.session_id)
         self.assertEqual(False, driver.w3c)
-
-    @httpretty.activate
-    def test_clipboard(self):
-        driver = TestHelper.mock_android_driver()
-        httpretty.register_uri(
-            httpretty.POST,
-            'http://localhost:4723/wd/hub/session/1234567890/appium/device/set_clipboard',
-            body='{"value": ""}'
-        )
-        driver.set_clipboard_text('hello')
-
-        d = json.loads(httpretty.last_request().body)
-        self.assertEqual("aGVsbG8=", d["content"])
-        self.assertEqual("plaintext", d["contentType"])
 
 
 if __name__ == "__main__":
